third_party/WebKit/LayoutTests/http/tests/security/xss-exception-expected.txt

CONSOLE MESSAGE: line 2: PASS: not executed in iframe context
CONSOLE MESSAGE: line 2: PASS: not executed in iframe context
CONSOLE MESSAGE: line 2: PASS: not executed in iframe context
Test passes if three "PASS" messages are logged to the console.  
